+++ CHANGELOG.md
@@ -6,18 +6,20 @@ also be mirrored to the Github releases page.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.1]
 
 ### Added
 
 - Callback function to mock sleep instruction.
 - Power module extra: `hal_power_change_module_powers`.
 - `hal_result_power` return type for the power module functions.
+- System module.
 
 ### Changed
 
 - Naming conventions for module members like functions.
 - C standard to C99.
+- System module's naming conventions.
 
 ### Fixed
 
@@ -91,7 +93,8 @@ branch.
   - TWI
   - ADC
 
-[unreleased]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.3...v0.4.0
 [0.3]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.2...v0.3
 [0.2]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.1...v0.2